backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
supabase>=2.3.0
httpx>=0.26.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
openai>=1.10.0
pandas>=2.1.0
numpy>=1.26.0
python-multipart>=0.0.6
apscheduler>=3.10.0
